Fix poor performance of holes
… by treating them the exact same as existential quantification.

The problem this is fixing is that code samples like
`./examples/reddit-haskell.grace` would take unusually long
to type-check.

The immediate problem was that the context kept getting larger
each time an expression was checked against a typed hole, which
slowed down all context-related operations.  The fix is to treat
typed holes exactly the same as existential quantification, by
cleaning up the context associated with them after their scope is
complete.