Merge `Syntax`/`Type` with their respective `Node` types
… and also label fields.

This simplifies the syntax tree and also makes it easier to
add more fields to constructors without disrupting the existing
code.

This is also conceptually more consistent since we may be
interested in locations other than at the beginning of constructors,
and it's weird to treat other types of locations differently.