Added Grace.Utils module
This module currently holds an orphan `Lift (Seq a)` instance that is
needed for containers < 0.6.6. Previously, this instance was taken from
the string-interpolate package, but it appears that this package doesn't
export the instance in the case of containers >= 0.6.6.
Anyways, providing our own seems to be more robust here.

Appart from that the module is meant to hold other fixes and utility
functions to provide a certain compatibility across GHC- and
library-versions.