Automatically insert `some`s
This is a more principled alternative to `T <: Optional T` (which was
removed in 94191d6d445332784d19b2963fbeb5bc7189ff21).  Instead of
making `T` a subtype of `Optional T` we automatically inject `some`
wherever there is a value whose type is expected to be `Optional T`
and then strip all `some`s at the end of interpretation.  This ensures
that we get the best of both worlds:

- `merge` can distinguish a "present `null`" (e.g.
  `List/head [ null ]`) from an "absent `null`" (e.g. `List/head [ ]`)

- the input and output of interpretation are still valid JSON