Merge import resolution and type inference
The motivation for this change is to make it so that `some` injection
(introduced in 658e20c6ef8a8aefed8484c403834c6bc2d8b1bd) works across
import boundaries.

In order for that to work we need to be able to push the `check`
judgment across import boundaries so that we can insert `some`s into
imported expressions.

Another upside of this change is that we no longer to need to add
special-case support for annotations on imports.  In fact, many imports
(including weakly-typed JSON imports) no longer need annotations,
period.