Fix type defaulting for prompt generation
Defaulting the type outside of `toJSONSchema` not only simplifies the
function, but also slightly improves handling the case where the user
supplies a `prompt` without a type annotation.  Before this change the
type annotation would have cased the `extractNonRecord` code path to be
hit (because the `schema` was not defaulted) but after this change we
correctly hit the `extractText` path.