Improve inference for a record as a subtype of JSON (#75)
The motivation for this change is so that something like this works:

```haskell
(http (GET{ url: "https://httpbin.org/get" })).headers
```

Before this change that would fail because `{ headers: JSON, a? }` is
not a subtype of `JSON`.  This change fixes that by adding a subtyping
rule to handle that case (defaulting `a?` to the empty set of fields).