Fixes to `Grace.Infer.check`
This bundles two fixes to `check` because I was too lazy to properly
split it into two commits.

The first fix is to fix support for `read "null": Optional T`.  Before
this change, that would get rewritten to `some (read "null" : T)` and
fail.  Now it will correctly decode `"null"`.

The second fix is to add `check` cases for record projections other
than `Single`, which improves the quality of type errors.