Add support for `prompt` with an incomplete schema
We now preserve type-checking state across evaluation, meaning that we
no longer need to default or quantify an incomplete type with unsolved
type variables.  This means that we can do stuff like this, which
previously would not work:

```haskell
let f = import prompt{ key, text: "A function to increment a Natural number" }

in { f, result: f 2 }
```

After this change, the type checker will infer an incomplete type with
type variables (e.g. `{ f: Natural -> a?, result: a? }`) and the LLM's
output will be checked against the same type (which will then be
further refined to `{ f: Natural -> Natural, result: Natural }`.

This would also work for other dynamic `import`s, too, but this just
implements it for `prompt` for now.