Thread type-checking state throughout evaluation
This updates the `Grace.{Normalize,Interpret}` functions to thread the
type-checking state so that the type-checking state can be *updated* by
evaluation.  This means that evaluation can lead to the inferred type
being refined, which comes in handy for the case of dynamic imports.

For example, after this change, the following works:

```haskell
>>> :type import prompt{ key: ../openai.key : Key, text: "Give me a function to increment a Natural number" }
Natural -> Natural
```

Previously that would have inferred `forall (a : Type) . a` as the
type, which would have been the wrong type.