Fix inferred types of imported values (#145)
In #143 I broke the inferred type of dynamically imported values
because when attempting to reset the `input` field of the `Status`
type I ended up resetting the entire type back to where it was
pre-import, which also reset the type-checking context.

This change fixes that by only modifying the `input` field.