Remove duplicate HTTP request for `import github`
The change in #143 introduced a duplicate HTTP request when using
`import github` because first `GitHub.github` would make an HTTP
request for the file (which would then be discarded) and then
`Interpret.interpretWith` would make another HTTP request for the same
file (because the `download_url` would be provided as the `input`),
which would be kept and used.

This change fixes `GitHub.github` to no longer perform the `http`
request at all.  Instead, it now just returns the `download_url` and
then if it is an `import` that will be passed to
`Interpret.interpretWith` as the `input` which will handle the HTTP
request.  If it is not an `import` then the non-`import` code path
will now handle the `http` request itself (the same as the original
`GitHub.github` code).