Fix test failures related to relative paths
Apparently CI is not running tests (probably my fault) and I will look
into that, but in the meantime I discovered that because tests were not
being run that two small test failures slipped through, which this
change fixes.

The most important fix is to strip `./` from relative paths before
making them relative to a URI.  The leading `./` was restored in #158
and just needed a matching change the `Semigroup` instance for `Input`.