Add support for simple pattern matching (#164)
Now you can use the `fold` keyword in such a way that it looks like traditional pattern matching, like this:

```haskell
let match = fold
        { Left number: show number
        , Right bool: if bool then "true" else "false"
        }

in  map match [ Left 2, Right false, Right true ] 
```

In fact, the reason this works is because it's not specific to the `fold` keyword.  The `fold` keyword *still* accepts an ordinary record of handlers, but what I did was to generalize record syntax to support function-like definitions of fields.  When you write a record like this:

```haskell
{ Left number: show number
, Right bool: if bool then "true" else "false"
}
```

… it gets desugared to:

```haskell
{ Left: \number -> show number
, Right: \bool -> if bool then "true" else "false"
}
```

In fact, records support almost the exact same syntax for defining functions as the `let` keyword.  The only difference is that a record uses `:` to separate the signature from the right-hand side instead of `=`.  This means you can do stuff like:

```haskell
{ Left (number: Natural): Text: show number
, Right (bool: Bool): Text: if bool then "true" else "false"
}
```