Fix destructuring on bind for `let`
This fixes the following example code:

```haskell
let { x: 0, y } = { x: 1, y: 2 }

in  x + y
```

The original issue was that the `2` in `y: 2` was being `check`ed
against the type `Optional Natural`, which means that it was being
elaborated to `some 2` before being bound to the `y` variable.

However, you can't add values of type `Optional Natural` which led to
the above code crashing.  There are a few possible ways to fix this,
such as:

- fixing the code to check the right-hand side (when present) against
  the non-`Optional` type (when present)

- undoing elaboration during the normalization phase

I *think* the former approach is the more principled and correct
approach, but I went with the latter approach because that's what the
code is already doing for non-record ("plain") destructured bindings.