Stop stripping `some`s (#183)
Back in https://github.com/Gabriella439/grace/commit/44a6871d79301fd9a9d9a81c3371036382647cc5
I implemented support for automatic `Optional` subtyping where the
type checker would insert missing `some`s so that downstream utilities
would work reliably during normalization.

However, as part of this change I also had the type-checker strip
the elaborated `some`s at the very end (when displaying the result to
the user) so that most users wouldn't even need to know that they
existed as an intermediate implementation detail of typechecking.

However, in retrospect I think this final step of stripping the `some`s
was actually a mistake and we should leave them in the final output
displayed to the user.  This can fix a lot of user confusion and
frustration.

As a concrete example, suppose the user runs something like this in the
REPL

```haskell
>>> [ 2 ].0  # Old behavior
2
```

Because the REPL is omitting the `some` in the displayed result the
user might then believe that they can extend the expression to add it
to another number, like this:

```haskell
>>> [ 2 ].0 + 2  # Type error
```

… which will fail because `[ 2 ].0` actually has type
`Optional Natural` and not `Natural`.

If we include the `some` in the result:

```haskell
>>> [ 2 ].0  # New behavior
some 2
```

… then it's more obvious to the user that the result cannot be added
as is and the `Optional` value needs to be handled in some way before
proceeding further.

More generally, we should try to educate users on how to use and
handle `some`s instead of trying to hide them from the user.

The one major exception to this rule is the `grace interpret` command,
where we still strip the `some`s because we want the output to be
JSON-compatible as much as possible.  Everywhere else we keep the
`some`s.