Overhaul dynamic import logic (#189)
The purpose of this commit is to overhaul the dynamic import logic so
that more interesting types can be inferred for non-`import` imports.
The simplest explanation of this change is:

```haskell
>>> read "1"  # Before
JSON
```

```haskell
>>> read "1"  # After
Natural
```

Before this change the latter would not have worked unless you used
`import read` instead of `read`, but the problem with that is that
using `import read` would trigger full-blown code
interpretation/evaluation which is undesirable in a lot of cases.  The
idea behind the distinction between `import` and non-`import` is that
the former is basically for people who want "eval" semantics and the
latter is for people who want non-eval / inert / plain JSON semantics.

However, fixing the latter to support inferrable types was tricky and
required a pretty significant overhaul of the code.  The main things
I did in service of this were:

- I changed `Grace.Infer` to always infer an unsolved variable if no
  schema was provided

  Previously it would prematurely a missing schema to a concrete type
  (e.g. `JSON` or `Text) which meant that more interesting types
  couldn't be inferred from the dynamic import's result.

- I changed `checkJSON` to handle unrecognized types by behaving
  very similar to `check`

  Basically, when it hits an unrecognized type (most notable: an
  unsolved type variable) it converts that JSON to Grace code and
  infers the type of that Grace code and also makes sure that the
  inferred type is a subset of JSON.

Along the way I also made some small changes to the semantics of the
various dynamic imports:

- `github` imports now default to `Text` by default

  This is *usually* what users want.  It's not going to be common to
  import JSON values from GitHub.  More likely the user will want to
  import files so that an LLM can read/browse them.

- `github`/`http` imports now special-case `Text` imports

  These now will import the contents raw (just like how ordinary
  imports behave with a `: Text` type annotation) rather than
  expecting a quoted string.