Fix parsing performance regression
In #163 I inadvertently introduced a parsing performance regression
where a series of `let` assignments of the form:

```
let x₀ = e₀
let x₁ = e₁
…
let xₙ = eₙ
in  …
```

… would lead to an exponential blow-up in the parser.  This is because
these sorts of `let` assignments could be parsed as either a
`Syntax.Definition` or a `Syntax.Bind` and the parser would parse both
possible alternatives in parallel for each assignment.  That in turn
meant that if you had N such assignments then the parser would be
parsing 2^N branches in parallel, leading to the exponential blowup.

The fix is to amend the parser to unambiguously parse these sorts of
assignments as `Sytnax.Definition`s and not `Syntax.Bind`s.