Automatically parallelize Grace side effects (#197)
In #139 (and #141) I added support for dynamic imports without schemas
and in the course of implementing that I had to undo Grace's support
for automatic parallelization at the time.  The reason why is because
it required threading type-checking state throughout evaluation and
that couldn't (easily) be parallelized.

Well, I finally got around to parallelizing it (which was indeed not
easy, but not as hard as I originally thought).  The trick was to
define a custom monad transformer (`Parallelizable`) which you can
think of as basically like `ConcurrentlyT` (the `Concurrently` type
generalized to a monad transformer).

The `Grace` monad (which I also defined in this change) is basically
`Parallelizable (StateT Status IO)` which means that it interleaves
concurrent segments with serial `StateT` segments and it takes care of
automatically parallelizing things when using `Applicative` operations
(including `ApplicativeDo` and `traverse`, most notably).

After updating the code to use this new `Grace` monad (and also
auditing the code to exploit `ApplicativeDo` and `traverse` as much as
possible to maximize concurrency) it "Just Works™️".  Multiple calls
to `prompt` (or any other import or side effect) will now automatically
run in parallel if they have no data dependencies between one another.