Fix type-checking of record definition syntax
The old code would break on the following example:

```
fold
  { some _: { a: A{ }, b: { } }
  , null:   { a: B{ } }
  }
```

… failing with a malformed context.  The fix is to type check
definitions using the same logic as type checking lambdas.  In other
words, instead of using `infer` we check against an unsolved variable.

This ensures that the inferred type of the lambda doesn't reference
any unsolved variables that go out of scope when the `scoped` blocks
end.