Improve fold-related error messages
The type-checker will no longer fall back to type checking a `fold`
as a union if the provided fields are a non-empty subset of a standard
fold type.  This should improve `fold`-related error messages, such as
in the following case:

```haskell
fold { zero: 0 }
```

Before that would fall back to treating the `fold` as a union and now
it recognizes that the user was attempting to `fold` a `Natural`
number and it will report the `succ` field as missing.