Add `export show` that can print *any* expression
`export show` is basically the dual of `import read`: it turns *any*
expression (after evaluation) into a string representing the
pretty-printed expression

Needless to say: this breaks parametricity, meaning that certain
classes of automatic program transformations are no longer necessarily
safe (not that we currently perform such program transformations but
breaking parametricity precludes us from doing so in the future).

You might argue that `import read : forall (a : Type) . Text -> a`
already broke parametricity, I guess.  Either way, parametricity is
**FUCKED**.

I personally think it's okay, though, and I think being able to convert
arbitrary expressions to `Text` is useful in its own right (such as
sending code for an LLM to try to understand).

However, the actual thing that tipped me over the edge towards
supporting this is that I view this part of a broader pattern of
special keywords that can `export` Grace code to various formats.

For example, you could imagine a `hash` keyword that by default
hashes `JSON` but when similarly prefixed with `export` (i.e.
`export hash`) is able to hash an arbitrary Grace expression.