Speed up type-checking of `fold` (#218)
I noticed that type checking was slowing down on the recently added
`examples/emotion-wheel.ffg` example.  When I dug in more closely I saw
that it was because the type checker was doing an excessive amount of
backtracking and that backtracking scales exponentially with how
many times you nest `fold`s.

The fix was to add a fast path for type-checking `fold`s which doesn't
backtrack at all.

One side-effect of this change is that Grace now rejects certain folds
that it used to accept.  See the changes to the golden test results for
more details.