trygrace.dev: Stop writing expression to query parameter
Before, trygrace.dev would record the expression to a query parameter
so that the URL would be sharable but in practice most useful
expressions were too large to share in this way so I'm deprecating
this feature.  The query parameter is still *read* for the value if
present (in order to support old URLs for backwards compatibility), but
the current expression is no longer written back to the query
parameter.  Instead, we track it via `sessionStorage` (tab-local
storage) so that it will still survive a tab refresh.