Change syntax for `Optional` comprehensions (#225)
The syntax for *both* `List` and `Optional` comprehensions used to be:

```haskell
for x of m in e
```

… which made it difficult to type check correctly and deteriorated the
quality of error messages.

In order to fix that, I've switched the syntax of `Optional`
comprehensions to be like Swift:

```haskell
if let x = m in e
```

For example:

```haskell
if let x = some 1 in x + 1  # some 2
```

This is also more efficient to type check, too, since we don't have to
backtrack.